ajustes nas rotas tipoBomba

+++ Tcc2/sistema/app/routes/tipoBomba.py
@@ -1,6 +1,6 @@
 from flask import render_template, url_for, redirect, request, flash 
 from app import app, db
-from app.models import TipoBomba, Peca, Bomba_peca
+from app.models import TipoBomba, Peca, Bomba_peca, OrdemServico
 from app.forms import RegistraTipoBombaForm
 from flask_login import login_required
 from sqlalchemy import text
@@ -10,12 +10,16 @@ from sqlalchemy import text
 def cadastroTipoBomba():
 
     form = RegistraTipoBombaForm()
-    form.rolamentoDianteiro.choices = [(peca.id, peca.descricao) for peca in Peca.query.filter_by(nome="Rolamento")]
-    form.rolamentoTraseiro.choices = [(peca.id, peca.descricao) for peca in Peca.query.filter_by(nome="Rolamento")]
-    form.retentorDianteiro.choices = [(peca.id, peca.descricao) for peca in Peca.query.filter_by(nome="Retentor")]
-    form.retentorTraseiro.choices = [(peca.id, peca.descricao) for peca in Peca.query.filter_by(nome="Retentor")]
-    form.tampaDianteira.choices = [(peca.id, peca.descricao) for peca in Peca.query.filter_by(nome="Tampa")]
-    form.tampaTraseira.choices = [(peca.id, peca.descricao) for peca in Peca.query.filter_by(nome="Tampa")]
+    rol = [(peca.id, peca.descricao) for peca in Peca.query.filter_by(nome="Rolamento")]
+    ret = [(peca.id, peca.descricao) for peca in Peca.query.filter_by(nome="Retentor")]
+    tamp = [(peca.id, peca.descricao) for peca in Peca.query.filter_by(nome="Tampa")]
+
+    form.rolamentoDianteiro.choices = rol
+    form.rolamentoTraseiro.choices = rol
+    form.retentorDianteiro.choices = ret
+    form.retentorTraseiro.choices = ret
+    form.tampaDianteira.choices = tamp
+    form.tampaTraseira.choices = tamp
     form.placa.choices = [(peca.id, peca.descricao) for peca in Peca.query.filter_by(nome="Placa")]
     form.eixo.choices = [(peca.id, peca.descricao) for peca in Peca.query.filter_by(nome="Eixo")]
     form.rotor.choices = [(peca.id, peca.descricao) for peca in Peca.query.filter_by(nome="Rotor")]
@@ -26,26 +30,30 @@ def cadastroTipoBomba():
         possuiTipobomba = TipoBomba.query.filter_by(tipo= form.tipo.data).all()
         
         if not possuiTipobomba or possuiTipobomba is None:
-        
-            listaPecas = [form.rolamentoDianteiro.data, form.rolamentoTraseiro.data,
-                        form.retentorDianteiro.data, form.retentorTraseiro.data,
-                        form.tampaDianteira.data, form.tampaTraseira.data,
-                        form.placa.data, form.eixo.data, form.rotor.data, form.bucha.data]
-            
-            tipobomba = TipoBomba(tipo=form.tipo.data, mca=form.mca.data, rotacao=form.rotacao.data)
-            db.session.add(tipobomba)
-            db.session.commit()
-            
-            getIdBomba = tipobomba.id
-            
-            for lista in listaPecas:
-                bb_peca = Bomba_peca(tipoBomba_id=getIdBomba, peca_id=lista)
-                db.session.add(bb_peca)
-                db.session.commit()   
-                
-            flash('Bomba cadastrada com sucesso!', 'info')
-            return redirect(url_for('listaTipoBombas'))
-        
+
+            try:
+                listaPecas = [form.rolamentoDianteiro.data, form.rolamentoTraseiro.data,
+                            form.retentorDianteiro.data, form.retentorTraseiro.data,
+                            form.tampaDianteira.data, form.tampaTraseira.data,
+                            form.placa.data, form.eixo.data, form.rotor.data, form.bucha.data]
+
+                tipobomba = TipoBomba(tipo=form.tipo.data, mca=form.mca.data, rotacao=form.rotacao.data)
+                db.session.add(tipobomba)
+                db.session.commit()
+
+                getIdBomba = tipobomba.id
+
+                for lista in listaPecas:
+                    bb_peca = Bomba_peca(tipoBomba_id=getIdBomba, peca_id=lista)
+                    db.session.add(bb_peca)
+                    db.session.commit()
+
+                flash('Bomba cadastrada com sucesso!', 'info')
+                return redirect(url_for('listaTipoBombas'))
+
+            except Exception as e:
+                print(e.args)
+
         flash("Já possui este Tipo/Modelo cadastrado!", 'error')
     
     return render_template('bomba/cadastro.html', form=form, icone="fas fa-plus", bloco1="Cadastro", bloco2="Bombas")
@@ -76,20 +84,26 @@ def editarTipoBomba(id):
         tipo = (request.form.get("tipo"))
         mca = (request.form.get("mca"))
         rotacao = (request.form.get("rotacao"))
+        qtEstoque = (request.form.get("qtEstoque"))
         
-        if tipo and mca and rotacao:
+        if tipo and mca and rotacao and qtEstoque:
             bomba = TipoBomba.query.filter_by(tipo=tipo).first()
 
             if not bomba or bomba is None or bomba.id == id:
-                tipoBomba.tipo = tipo
-                tipoBomba.mca = mca
-                tipoBomba.rotacao = rotacao
 
-                db.session.commit()
+                try:
+                    tipoBomba.tipo = tipo
+                    tipoBomba.mca = mca
+                    tipoBomba.rotacao = rotacao
+                    tipoBomba.qtEstoque = qtEstoque
 
-                flash('Salvo com sucesso!', 'info')
-                return redirect(url_for("listaTipoBombas"))
+                    db.session.commit()
 
+                    flash('Salvo com sucesso!', 'info')
+                    return redirect(url_for("listaTipoBombas"))
+
+                except Exception as e:
+                    print(e.args)
             flash('Já possui este Tipo/Modelo cadastrado!', 'error')
 
     return render_template("bomba/editar.html", tipoBomba = tipoBomba, icone="fas fa-pen", bloco1="Edição", bloco2="Bombas")
@@ -100,17 +114,21 @@ def editarTipoBomba(id):
 def excluirTipoBomba(id):
     bomba = TipoBomba.query.filter_by(id=id).first()
     bb_peca = Bomba_peca.query.filter_by(tipoBomba_id=id).all()
+    ordemServico = OrdemServico.query.filter_by(equipamento=id).first()
 
-    if bomba:
-        for p in bb_peca:
-            db.session.delete(p)
+    if bomba and ordemServico is None:
+        try:
+            for p in bb_peca:
+                db.session.delete(p)
 
-        db.session.delete(bomba)
-        db.session.commit()
+            db.session.delete(bomba)
+            db.session.commit()
 
+            flash("Bomba e excluida com sucesso!", 'info')
+            return redirect(url_for('listaTipoBombas'))
 
-        flash("Bomba e excluida com sucesso!", 'info')
-        return redirect(url_for('listaTipoBombas'))
+        except Exception as e:
+            print(e.args)
 
     flash("Não é possível excluir!", 'error')
     return redirect(url_for('listaTipoBombas'))
